+++ eugene-web/src/webapp/templates/base.hbs
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <title>Eugene</title>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <meta name="theme-color" content="#ffffff">
+    <script src="https://unpkg.com/htmx.org@1.9.12"></script>
+</head>
+
+<body>
+    <div id="body-container">
+        <h1>Eugene</h1>
+{{> body }}
+    </div>
+</body>
+</html>
