+++ eugene-web/src/webapp/templates/index.hbs
@@ -1,0 +1,23 @@
+{{#* inline "body" }}
+    <h2>Enter some scripts</h2>
+    <form hx-post="/eugene/app/lint.html" hx-target="#result">
+    <label for="sql">SQL</label>
+    <textarea id="sql" name="sql" rows="20" cols="100">
+-- file: create_table.sql
+create table books (
+    id serial primary key,
+    title text,
+    author text,
+    published date
+);
+-- file: alter_table.sql
+alter table books
+  alter column text set not null;
+alter table books
+  alter column author set not null;
+    </textarea>
+    <button type="submit">Lint</button>
+    </form>
+    <div id="result"></div>
+{{/inline}}
+{{> base }}
